[Feature #162450345] moved validation to validations folder + moved test incidents data to separate file
