+++ app/api/v2/views/auth.py
@@ -2,7 +2,7 @@ import json
 from flask import request, json
 from flask_restful import Resource
 from app.api.v2.models.users_model import UsersModel
-from app.api.v2.views.validation import ViewsValidation
+from app.api.v2.validation.validation import ViewsValidation
 from app.api.v2.auth_decorators.auth_decorator import isAdmin
