+++ app/test/v2/views/test_incidents_data.py
@@ -1,0 +1,66 @@
+# test data
+incident1 = {
+    'title': 'Traffice Corruption',
+    'description': 'Offices taking bribes',
+    'type': 'RedFlag',
+    'incident_status': 'Rejected',
+    'location': '-34444400, 3444499900',
+    'image': [
+        {
+            'dir': 'var/www/uploads/incidents/img/USR-232455.jpeg',
+            'filesize': '2045kb'
+        },
+        {
+            'dir': 'var/www/uploads/incidents/img/USR-232455.jpeg',
+            'filesize': '2045kb'
+        }
+    ],
+    'video': [
+        {
+            'dir': 'var/www/uploads/readflags/video/USR-232455.mp4',
+            'filesize': '340098245Kb'
+        }, {
+            'dir': 'var/www/uploads/readflags/video/USR-232455.mp4',
+            'filesize': '340098245Kb'
+        }
+    ],
+    'comment': 'This is fake all news i tell ya'
+
+
+}
+# test data
+
+incident2 = {
+    "title": "Manhole Hazard",
+    "description": "Manhole open around hirlingum mall",
+    "type": "Intervention",
+    "incident_status": "Rejected",
+    "location": "-34444400, 3444499900",
+    "image": [
+        {
+            "dir": "var/www/uploads/incidents/img/USR-232455.jpeg",
+            "filesize": "2045kb"
+        },
+        {
+            "dir": "var/www/uploads/incidents/img/USR-232455.jpeg",
+            "filesize": "2045kb"
+        }
+    ],
+    "video": [
+        {
+            "dir": "var/www/uploads/readflags/video/USR-232455.mp4",
+            "filesize": "340098245Kb"
+        }, {
+            "dir": "var/www/uploads/readflags/video/USR-232455.mp4",
+            "filesize": "340098245Kb"
+        }
+    ],
+    "comment": "ignore it or cover it up, out of sight out of mind"
+}
+
+comment_data = {
+    "comment": "Scintilating"
+}
+location_data = {
+    "location": "-3455666, 42254555"
+}
